+++ udata/commands/__init__.py
@@ -72,7 +72,7 @@ def register_commands(manager):
         try:
             __import__(name)
         except ImportError as e:
-            log.warning('Error importing %s: %s', name, e)
+            pass
         except Exception as e:
             log.error('Error during import of %s: %s', name, e)